feat: create ExternalLibraryService interface, rename existing class to DefaultExternalLibraryService, add getShellClassLoader method as it was in engine; remove not used librariesUpdateGetter
