+++ micro-engine/src/test/java/org/qubership/integration/platform/engine/consul/updates/UpdateGetterProducerTest.java
@@ -81,31 +81,6 @@ class UpdateGetterProducerTest {
     }
 
     @Test
-    void shouldCreateLibrariesUpdateGetterWithConfiguredKeyAndParser() {
-        UpdateGetterHelper<List<CompiledLibraryUpdate>> getter =
-                producer.librariesUpdateGetter(() -> consulClient, librariesUpdateParser);
-
-        List<KeyValue> entries = List.of(mock(KeyValue.class));
-        KeyValueList kvList = changedKvList(entries);
-        List<CompiledLibraryUpdate> updates = List.of(
-                mock(CompiledLibraryUpdate.class),
-                mock(CompiledLibraryUpdate.class)
-        );
-
-        when(consulClient.getValuesWithOptions(
-                eq("config/test/qip-engine-configurations/libraries-update"),
-                any(BlockingQueryOptions.class)
-        )).thenReturn(Uni.createFrom().item(kvList));
-        when(librariesUpdateParser.apply(entries)).thenReturn(updates);
-
-        AtomicReference<List<CompiledLibraryUpdate>> result = new AtomicReference<>();
-
-        getter.checkForUpdates(result::set);
-
-        assertSame(updates, result.get());
-    }
-
-    @Test
     void shouldCreateChainRuntimePropertiesUpdateGetterWithConfiguredKeyAndParser() {
         UpdateGetterHelper<Map<String, ChainRuntimeProperties>> getter =
                 producer.chainRuntimePropertiesUpdateGetter(() -> consulClient, chainRuntimePropertiesUpdateParser);
